+++ shuup_tests/utils/test_analog.py
@@ -15,20 +15,9 @@ def test_analog():
     PseudoPaymentProcessorLogEntry = define_log_model(PseudoPaymentProcessor)
     assert PseudoPaymentProcessorLogEntry.__module__ == PseudoPaymentProcessor.__module__
 
-    related_field_name = "related"
-    # Behavior changs in Django 1.9
-    if VERSION >= (1, 9):
-        related_field_name = "rel"
-    
-    if VERSION > (2, 0):
-        relation_manager = getattr(PseudoPaymentProcessorLogEntry._meta.get_field("target"), "remote_field")
-    else:
-        relation_manager = getattr(PseudoPaymentProcessorLogEntry._meta.get_field("target"), related_field_name)
-    
-    if VERSION < (2, 0):   
-        assert relation_manager.to is PseudoPaymentProcessor
-    else:
-        assert relation_manager.model is PseudoPaymentProcessor
+    related_field_name = "rel"
+    relation_manager = getattr(PseudoPaymentProcessorLogEntry._meta.get_field("target"), "remote_field")
+    assert relation_manager.model is PseudoPaymentProcessor
 
     relation_manager = getattr(PseudoPaymentProcessor.log_entries, "rel")
     assert relation_manager.model is PseudoPaymentProcessor