+++ shuup_workbench/test_urls.py
@@ -4,7 +4,6 @@
 #
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
-import django
 from django.conf import settings
 from django.conf.urls import include, url
 from django.conf.urls.static import static
@@ -13,16 +12,8 @@ from django.contrib import admin
 """
 Modify these only for Shuup tests. For testing modify urls.py instead.
 """
-if django.VERSION < (2, 0):
-    urlpatterns = [
-        url(r'^admin/', include(admin.site.urls)),
-        url(r'^sa/', include('shuup.admin.urls', namespace="shuup_admin", app_name="shuup_admin")),
-        url(r'^', include('shuup.front.urls', namespace="shuup", app_name="shuup")),
-    ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
-else:
-    from django.urls import path
-    urlpatterns = [
-        path('admin/', admin.site.urls),
-        url(r'^sa/', include('shuup.admin.urls', namespace="shuup_admin")),
-        url(r'^', include('shuup.front.urls', namespace="shuup")),
-    ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+urlpatterns = [
+    url(r'^admin/', include(admin.site.urls)),
+    url(r'^sa/', include('shuup.admin.urls', namespace="shuup_admin", app_name="shuup_admin")),
+    url(r'^', include('shuup.front.urls', namespace="shuup", app_name="shuup")),
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)